Added support for transparent activation and persistence to Proteomik API, Central Lookup now supports LookupListener registration for speicifc types. Keeps only single reference to identical/equal objects. Added PropertySheet support for GelNode. Modified ProteomicProjectFactory and GelViewer to create/accept relative paths to gels. Allows easy relocation of projects.

+++ Proteomik Project/test/unit/src/de/unibielefeld/gi/kotte/laborprogramm/project/ProjectTest.java
@@ -57,6 +57,9 @@ import static org.junit.Assert.*;
  */
 public class ProjectTest extends TestCase {
 
+    private File tmpDir;
+    private File tmpProjectDir;
+
     public ProjectTest() {
     }
 
@@ -68,34 +71,55 @@ public class ProjectTest extends TestCase {
     public static void tearDownClass() throws Exception {
     }
 
-    public static junit.framework.Test suite() {
-        return NbModuleSuite.create(NbModuleSuite.createConfiguration(ProjectTest.class));
+    @Override
+    protected void setUp() throws Exception {
+        tmpDir = new File(System.getProperty("java.io.tmpdir"));
+        tmpProjectDir = new File(tmpDir, "testDB4oProject");
     }
 
+    @Override
+    protected void tearDown() throws Exception {
+        deleteChildren(tmpProjectDir);
+        tmpProjectDir.delete();
+    }
+
+    public static junit.framework.Test suite() {
+        return NbModuleSuite.create(NbModuleSuite.createConfiguration(
+                ProjectTest.class));
+    }
+
+    /**
+     * Test plain db4o persistence. No Netbeans related APIs used here.
+     *
+     */
     @Test
     public void testDB4oPersistence() {
-        File tmpDir = new File(System.getProperty("java.io.tmpdir"));
-        File tmpProjectDir = new File(tmpDir, "testDB4oProject");
-        if(tmpProjectDir.exists()) {
+
+        if (tmpProjectDir.exists()) {
             deleteChildren(tmpProjectDir);
         }
-        File databaseFile = new File(tmpProjectDir,"plopp.ppr");
-        //tmpProjectDir.deleteOnExit();
+        File databaseFile = new File(tmpProjectDir, "plopp.ppr");
         tmpProjectDir.mkdirs();
         ICrudProvider icp;
         IProject ipr1 = null;
+        //create seesion
         try {
-            icp = Lookup.getDefault().lookup(ICrudProviderFactory.class).getCrudProvider(databaseFile.toURI().toURL(), new NoAuthCredentials(),Thread.currentThread().getContextClassLoader());
+            icp = Lookup.getDefault().lookup(ICrudProviderFactory.class).
+                    getCrudProvider(databaseFile.toURI().toURL(),
+                    new NoAuthCredentials(), Lookup.getDefault().lookup(
+                    ClassLoader.class));
             icp.open();
             ICrudSession ics = icp.createSession();
             ics.open();
             ipr1 = createTestProject();
+            //create object in database
             ics.create(Arrays.asList(ipr1));
             //test retrieval in same session
             Collection<IProject> coll = ics.retrieve(IProject.class);
-            assertEquals(1,coll.size());
+            assertEquals(1, coll.size());
             IProject ipr = coll.iterator().next();
             assertNotNull(ipr);
+            //close session
             ics.close();
             icp.close();
         } catch (MalformedURLException ex) {
@@ -104,33 +128,90 @@ public class ProjectTest extends TestCase {
 
         //test retrieval in new session
         try {
-            icp = Lookup.getDefault().lookup(ICrudProviderFactory.class).getCrudProvider(databaseFile.toURI().toURL(), new NoAuthCredentials(),Thread.currentThread().getContextClassLoader());
+            icp = Lookup.getDefault().lookup(ICrudProviderFactory.class).
+                    getCrudProvider(databaseFile.toURI().toURL(),
+                    new NoAuthCredentials(), Thread.currentThread().
+                    getContextClassLoader());
             icp.open();
             ICrudSession ics = icp.createSession();
             ics.open();
             IProject ipr2 = ics.retrieve(IProject.class).iterator().next();
             assertNotNull(ipr2);
-            assertEquals(ipr1.toString(),ipr2.toString());
+            assertEquals(ipr1.toString(), ipr2.toString());
+            ics.close();
+            icp.close();
+        } catch (MalformedURLException ex) {
+            Exceptions.printStackTrace(ex);
+        }
+    }
+
+    /**
+     * Test modification after first persistence to db was performed.
+     */
+    @Test
+    public void testProjectCRUDCapability() {
+        if (tmpProjectDir.exists()) {
+            deleteChildren(tmpProjectDir);
+        }
+        File databaseFile = new File(tmpProjectDir, "plopp.ppr");
+        //tmpProjectDir.deleteOnExit();
+        tmpProjectDir.mkdirs();
+        ICrudProvider icp;
+        IProject ipr1 = null;
+        try {
+            icp = Lookup.getDefault().lookup(ICrudProviderFactory.class).
+                    getCrudProvider(databaseFile.toURI().toURL(),
+                    new NoAuthCredentials(), Thread.currentThread().
+                    getContextClassLoader());
+            icp.open();
+            ICrudSession ics = icp.createSession();
+            ics.open();
+            ipr1 = createTestProject();
+            ics.create(Arrays.asList(ipr1));
+            //test retrieval in same session
+            Collection<IProject> coll = ics.retrieve(IProject.class);
+            assertEquals(1, coll.size());
+            IProject ipr = coll.iterator().next();
+            assertNotNull(ipr);
+
+            //test update of original object
+
+            ipr = addSpotsAndWells(ipr);
+            ics.update(ipr);
             ics.close();
             icp.close();
         } catch (MalformedURLException ex) {
             Exceptions.printStackTrace(ex);
         }
 
-        deleteChildren(tmpProjectDir);
-        tmpProjectDir.delete();
+        //test retrieval in new session
+        try {
+            icp = Lookup.getDefault().lookup(ICrudProviderFactory.class).
+                    getCrudProvider(databaseFile.toURI().toURL(),
+                    new NoAuthCredentials(), Thread.currentThread().
+                    getContextClassLoader());
+            icp.open();
+            ICrudSession ics = icp.createSession();
+            ics.open();
+            IProject ipr2 = ics.retrieve(IProject.class).iterator().next();
+            assertNotNull(ipr2);
+            assertFalse(ipr2.get96Plates().isEmpty());
+            System.out.println("IProject has the following plates: "+ipr2.get96Plates());
+            ics.close();
+            icp.close();
+        } catch (MalformedURLException ex) {
+            Exceptions.printStackTrace(ex);
+        }
     }
 
-    //@Test
+    @Test
     public void testProject() {
         //set up project in test directory
         //IProteomicProjectFactory ippf = Lookup.getDefault().lookup(IProteomicProjectFactory.class);
         ProjectFactory ippf = new ProteomikProjectFactory();
         assertNotNull(ippf);
         IProject testProject = createTestProject();
-        File tmpDir = new File(System.getProperty("java.io.tmpdir"));
-        File tmpProjectDir = new File(tmpDir, "testProject");
-        if(tmpProjectDir.exists()) {
+        if (tmpProjectDir.exists()) {
             deleteChildren(tmpProjectDir);
         }
         //tmpProjectDir.deleteOnExit();
@@ -139,7 +220,8 @@ public class ProjectTest extends TestCase {
         IProteomicProject ipp = ippf2.createProject(tmpProjectDir, testProject);
         String str1 = null;
         try {
-            ipp = (IProteomicProject) ippf.loadProject(FileUtil.toFileObject(tmpProjectDir), new ProjectState() {
+            ipp = (IProteomicProject) ippf.loadProject(FileUtil.toFileObject(
+                    tmpProjectDir), new ProjectState() {
 
                 @Override
                 public void markModified() {
@@ -156,25 +238,35 @@ public class ProjectTest extends TestCase {
         }
 //        try {
 //            ipp.activate(new File(tmpProjectDir, ProteomikProjectFactory.PROJECT_FILE).toURI().toURL());
-            str1 = ipp.toString();
-            ipp.close();
+        str1 = ipp.toString();
+        ipp.close();
 //        } catch (MalformedURLException ex) {
 //            Exceptions.printStackTrace(ex);
 //        }
-        DB4oCrudProvider instance = new DB4oCrudProvider(new File(tmpProjectDir,ProteomikProjectFactory.PROJECT_FILE), new NoAuthCredentials(),Thread.currentThread().getContextClassLoader());
+
+        //test retrieval from database
+
+        DB4oCrudProvider instance = new DB4oCrudProvider(new File(tmpProjectDir,
+                ProteomikProjectFactory.PROJECT_FILE), new NoAuthCredentials(), Lookup.
+                getDefault().lookup(ClassLoader.class));
         instance.open();
         ICrudSession ics = instance.createSession();
         ics.open();
         Collection<IProject> projects = ics.retrieve(IProject.class);
-        Logger.getLogger(getClass().getName()).log(Level.INFO,"Retrieved {0} projects",projects.size());
+        Logger.getLogger(getClass().getName()).log(Level.INFO,
+                "Retrieved {0} projects", projects.size());
         assertEquals(1, projects.size());
-        for(IProject proj:projects) {
-            Logger.getLogger(getClass().getName()).log(Level.INFO,"Retrieved project {0}",proj);
+        for (IProject proj : projects) {
+            Logger.getLogger(getClass().getName()).log(Level.INFO,
+                    "Retrieved project {0}", proj);
         }
         ics.close();
         instance.close();
+
+        //retrieve from database and compare to previous version
         try {
-            IProteomicProject ipp2 = (IProteomicProject) ippf.loadProject(FileUtil.toFileObject(tmpProjectDir), new ProjectState() {
+            IProteomicProject ipp2 = (IProteomicProject) ippf.loadProject(FileUtil.
+                    toFileObject(tmpProjectDir), new ProjectState() {
 
                 @Override
                 public void markModified() {
@@ -186,65 +278,73 @@ public class ProjectTest extends TestCase {
                     //throw new UnsupportedOperationException("Not supported yet.");
                 }
             });
-            System.err.println("Str1: "+str1);
+            System.err.println("Str1: " + str1);
             String str2 = ipp2.toString();
-            System.err.println("Str2: "+str2);
+            System.err.println("Str2: " + str2);
             assertEquals(str1, str2);
             ipp2.close();
             //System.out.println("Project2: "+ipp2.toString());
         } catch (IOException ex) {
             Exceptions.printStackTrace(ex);
         }
-
-        deleteChildren(tmpProjectDir);
-        tmpProjectDir.delete();
         //IProteomicProject pp = ippf.createProject(dir);
     }
 
     private void deleteChildren(File f) {
-        File[] files = f.listFiles();
-        for (File file : files) {
-            if (file.isDirectory()) {
-                deleteChildren(f);
-            } else {
-                file.delete();
+        if (f != null) {
+            File[] files = f.listFiles();
+            for (File file : files) {
+                if (file.isDirectory()) {
+                    deleteChildren(f);
+                } else {
+                    file.delete();
+                }
             }
         }
     }
 
     private IProject createTestProject() {
         IProject project = null;
-        project = Lookup.getDefault().lookup(IProjectFactory.class).createEmptyProject();
+        project = Lookup.getDefault().lookup(IProjectFactory.class).
+                createEmptyProject();
         project.setName("Dummy Projekt");
-        project.setDescription("Testprojekt zum testen der Proteomik API Datenstrukturen");
+        project.setDescription(
+                "Testprojekt zum testen der Proteomik API Datenstrukturen");
         project.setOwner("kotte");
 
         //set up logical gel group
-        ILogicalGelGroup logical = Lookup.getDefault().lookup(ILogicalGelGroupFactory.class).createLogicalGelGroup();
+        ILogicalGelGroup logical = Lookup.getDefault().lookup(
+                ILogicalGelGroupFactory.class).createLogicalGelGroup();
         logical.setName("Testgruppe high level");
         logical.setDescription("Mock Objekt zum Test der Datenstruktur");
         logical.setParent(project);
         project.addGelGroup(logical);
 
         //set up two bio rep gel groups
-        IBioRepGelGroup biorep1 = Lookup.getDefault().lookup(IBioRepGelGroupFactory.class).createBioRepGelGroup();
+        IBioRepGelGroup biorep1 = Lookup.getDefault().lookup(
+                IBioRepGelGroupFactory.class).createBioRepGelGroup();
         biorep1.setName("Testgruppe #1 mid level");
         biorep1.setDescription("Mock Objekt zum Test der Datenstruktur");
         biorep1.setParent(logical);
         logical.addGelGroup(biorep1);
-        IBioRepGelGroup biorep2 = Lookup.getDefault().lookup(IBioRepGelGroupFactory.class).createBioRepGelGroup();
+
+        IBioRepGelGroup biorep2 = Lookup.getDefault().lookup(
+                IBioRepGelGroupFactory.class).createBioRepGelGroup();
         biorep2.setName("Testgruppe #2 mid level");
         biorep2.setDescription("Mock Objekt zum Test der Datenstruktur");
         biorep2.setParent(logical);
         logical.addGelGroup(biorep2);
 
         //set up two tech rep gel groups (one for each bio rep gel group)
-        ITechRepGelGroup techrep1 = Lookup.getDefault().lookup(ITechRepGelGroupFactory.class).createTechRepGelGroup();
+        ITechRepGelGroup techrep1 = Lookup.getDefault().lookup(
+                ITechRepGelGroupFactory.class).createTechRepGelGroup();
         techrep1.setName("Testgruppe #1 low level");
         techrep1.setDescription("Mock Objekt zum Test der Datenstruktur");
         techrep1.setParent(biorep1);
         biorep1.addGelGroup(techrep1);
-        ITechRepGelGroup techrep2 = Lookup.getDefault().lookup(ITechRepGelGroupFactory.class).createTechRepGelGroup();
+
+        ITechRepGelGroup techrep2 = Lookup.getDefault().lookup(
+                ITechRepGelGroupFactory.class).createTechRepGelGroup();
         techrep2.setName("Testgruppe #2 low level");
         techrep2.setDescription("Mock Objekt zum Test der Datenstruktur");
         techrep2.setParent(biorep2);
@@ -256,11 +356,24 @@ public class ProjectTest extends TestCase {
         gel1.setName("Testgel #1");
         gel1.setParent(techrep1);
         techrep1.addGel(gel1);
+
         IGel gel2 = Lookup.getDefault().lookup(IGelFactory.class).createGel();
         gel2.setDescription("Mock Objekt zum Test der Datenstruktur");
         gel2.setName("Testgel #2");
         gel2.setParent(techrep2);
         techrep2.addGel(gel2);
+
+        return project;
+    }
+
+    private IProject addSpotsAndWells(IProject project) {
+        //set up a 96 well microplate
+        IPlate96 plate96 = Lookup.getDefault().lookup(IPlate96Factory.class).
+                createPlate96();
+        plate96.setName("Testplatte");
+        plate96.setDescription("Mock Objekt zum Test der Datenstruktur");
+        plate96.setParent(project);
+        project.add96Plate(plate96);
 
         //set up some spots for the two gels
         ISpot spot11 = Lookup.getDefault().lookup(ISpotFactory.class).createSpot();
@@ -268,6 +381,8 @@ public class ProjectTest extends TestCase {
         spot11.setPosX(100);
         spot11.setPosY(199);
         spot11.setLabel("Spot #1 on gel #1");
+        IGel gel1 = project.getGelGroups().get(0).getGelGroups().get(0).
+                getGelGroups().get(0).getGels().get(0);
         spot11.setGel(gel1);
         gel1.addSpot(spot11);
         ISpot spot12 = Lookup.getDefault().lookup(ISpotFactory.class).createSpot();
@@ -277,11 +392,14 @@ public class ProjectTest extends TestCase {
         spot12.setLabel("Spot #2 on gel #1");
         spot12.setGel(gel1);
         gel1.addSpot(spot12);
+
         ISpot spot21 = Lookup.getDefault().lookup(ISpotFactory.class).createSpot();
         spot21.setNumber(1);
         spot21.setPosX(99);
         spot21.setPosY(200);
         spot21.setLabel("Spot #1 on gel #2");
+        IGel gel2 = project.getGelGroups().get(0).getGelGroups().get(0).
+                getGelGroups().get(0).getGels().get(0);
         spot21.setGel(gel2);
         gel2.addSpot(spot21);
         ISpot spot22 = Lookup.getDefault().lookup(ISpotFactory.class).createSpot();
@@ -293,7 +411,8 @@ public class ProjectTest extends TestCase {
         gel2.addSpot(spot22);
 
         //set up groups of spots
-        ISpotGroup spotgroup1 = Lookup.getDefault().lookup(ISpotGroupFactory.class).createSpotGroup();
+        ISpotGroup spotgroup1 = Lookup.getDefault().lookup(
+                ISpotGroupFactory.class).createSpotGroup();
         spotgroup1.setLabel("Spot #1");
         spotgroup1.setNumber(1);
         spotgroup1.addSpot(spot11);
@@ -301,7 +420,8 @@ public class ProjectTest extends TestCase {
         spotgroup1.addSpot(spot21);
         spot21.setGroup(spotgroup1);
         project.addSpotGroup(spotgroup1);
-        ISpotGroup spotgroup2 = Lookup.getDefault().lookup(ISpotGroupFactory.class).createSpotGroup();
+        ISpotGroup spotgroup2 = Lookup.getDefault().lookup(
+                ISpotGroupFactory.class).createSpotGroup();
         spotgroup2.setLabel("Spot #2");
         spotgroup2.setNumber(2);
         spotgroup2.addSpot(spot12);
@@ -309,13 +429,6 @@ public class ProjectTest extends TestCase {
         spotgroup2.addSpot(spot22);
         spot22.setGroup(spotgroup2);
         project.addSpotGroup(spotgroup2);
-
-        //set up a 96 well microplate
-        IPlate96 plate96 = Lookup.getDefault().lookup(IPlate96Factory.class).createPlate96();
-        plate96.setName("Testplatte");
-        plate96.setDescription("Mock Objekt zum Test der Datenstruktur");
-        plate96.setParent(project);
-        project.add96Plate(plate96);
 
         //collect some wells for our spots from the plates
         IWell96 wellA1 = plate96.getWell('A', 1);
@@ -342,7 +455,8 @@ public class ProjectTest extends TestCase {
         spot22.setStatus(SpotStatus.PICKED);
 
         //create 384 well plate
-        IPlate384 plate384 = Lookup.getDefault().lookup(IPlate384Factory.class).createPlate384();
+        IPlate384 plate384 = Lookup.getDefault().lookup(IPlate384Factory.class).
+                createPlate384();
         plate384.setParent(project);
         project.add384Plate(plate384);
         plate384.setName("grosse Testplatte");
